requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
#lightgbm==4.3.0
xgboost==2.0.3
optuna==3.6.1
matplotlib==3.8.4
seaborn==0.13.2
jupyter==1.0.0
kaggle==1.6.12
pyarrow==15.0.2
fastparquet==2024.2.0
python-dotenv==1.0.1
mlflow==2.11.3
